NODE-338: Moved GrpcServer in a new comm.grpc package.
